Only change directory if one has been specified
Previously, with-ftp would always change the working directory, even
if the path component of the URL was empty or "/".  The first slash
after the host name needs to be considered a separator and not part of
the directory name of the remote, though, and the urly library always
normalizes URL strings contain at least "/" as path.  With this
change, with-ftp will only attempt to change the directory if a path
component is present in the URL and it is not "/".  It will also strip
the first slash to make the path relative to the current directly.

Thus:

ftp://example.com => no chdir
ftp://example.com/ => no chdir
ftp://example.com/foo => chdir to "foo"
ftp://example.com//foo => chdir to "/foo"