Don't double-decode URL user-info; fixes #21
The URL's user-info component was being decoded once by `urly`, and then
subsequently again by `with-ftp`.  If the username or password contains
a '%' character, the second decode will fail or produce the wrong result.

The commit removes `urly`, since it didn't seem to be adding much, and
still the bug is not properly fixable while still using it: `urly`
decodes the user-info before returning it, but this leads to ambiguity
if the (decoded) username contains a ':' (unlikely, but legal).